BJS2-96836 refactor
